refactor: moved the logic to main layout
